update: 세션 관련 앱 분리
