+++ session/urls.py
@@ -1,10 +1,10 @@
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 
-from program.viewsets import ProposalViewSet
+from session.viewsets import SessionViewSet
 
 session_router = DefaultRouter()
-session_router.register("", ProposalViewSet, basename="session")
+session_router.register("", SessionViewSet, basename="session")
 
 urlpatterns = [
     path("", include(session_router.urls)),
