+++ session/viewsets.py
@@ -1,17 +1,17 @@
 from rest_framework.permissions import IsAuthenticatedOrReadOnly
 from rest_framework.viewsets import ModelViewSet
 
-from program.models import Proposal
-from program.serializers import ProposalListSerializer, ProposalSerializer
+from session.models import Session
+from session.serializers import SessionListSerializer, SessionSerializer
 
 
-class ProposalViewSet(ModelViewSet):
-    queryset = Proposal.objects.all()
+class SessionViewSet(ModelViewSet):
+    queryset = Session.objects.all()
     permission_classes = [IsAuthenticatedOrReadOnly]
     http_method_names = ["get", "head", "options"]
 
     def get_serializer_class(self):
         if self.action == "list":
-            return ProposalListSerializer
+            return SessionListSerializer
         else:
-            return ProposalSerializer
+            return SessionSerializer
